reverted changes from HistoryManager
